backend/requirements.txt
fastapi
uvicorn
pillow
python-multipart
numpy